rename: indext.tsx 타입 register 폴더로 이동 및 path 적용
