--- src/types/index.ts
@@ -1,15 +1,0 @@
-// FormButton component props
-export interface ButtonProps {
-  label?: string;
-  color?: string;
-  backgroundColor?: string;
-  borderColor?: string;
-  onClick?: () => void;
-}
-
-// SnsCircle
-export interface CircleProps {
-  color?: string;
-  marginRight?: number;
-  marginLeft?: number;
-}
