+++ src/types/register/registerPage.ts
@@ -51,3 +51,19 @@ export interface InputProps {
   subText?: string | ReactJSXElement;
   type?: string;
 }
+
+// FormButton component props
+export interface ButtonProps {
+  label?: string;
+  color?: string;
+  backgroundColor?: string;
+  borderColor?: string;
+  onClick?: () => void;
+}
+
+// SnsCircle
+export interface CircleProps {
+  color?: string;
+  marginRight?: number;
+  marginLeft?: number;
+}
